Support any category - Convert the current Plugin.Category as  Plugin.Category.Default - Present Plugin.Category parametered cases - Fetch all plugins with category property and display in sidebar
