+++ App/Shared/Data/Plugin+SwiftUI.swift
@@ -10,28 +10,37 @@ import Foundation
 extension Plugin.Category {
     var icon: String {
         switch self {
-        case .cache:
-            "tray"
-        case .controls:
-            "dpad"
-        case .draw:
-            "scribble"
-        case .highlighter:
-            "target"
-        case .info:
-            "info"
-        case .layer:
-            "square.3.layers.3d"
-        case .mapTiles:
-            "map"
-        case .portalInfo:
-            "selection.pin.in.out"
-        case .tweaks:
-            "wrench.and.screwdriver"
-        case .misc:
-            "ellipsis"
-        case .debug:
-            "ladybug"
+        case .default(let value):
+            switch value {
+            case .cache:
+                "tray"
+            case .controls:
+                "dpad"
+            case .draw:
+                "scribble"
+            case .highlighter:
+                "target"
+            case .info:
+                "info"
+            case .layer:
+                "square.3.layers.3d"
+            case .mapTiles:
+                "map"
+            case .portalInfo:
+                "selection.pin.in.out"
+            case .tweaks:
+                "wrench.and.screwdriver"
+            case .misc:
+                "ellipsis"
+            case .debug:
+                "ladybug"
+            }
+        default:
+            "puzzlepiece.extension"
         }
     }
+}
+
+extension Plugin.Category : Hashable {
+    
 }
